# app/web/openapi.json
{
  "components": {
    "schemas": {
      "ActionRequest": {
        "properties": {
          "action_id": {
            "title": "Action Id",
            "type": "integer"
          },
          "client_id": {
            "title": "Client Id",
            "type": "string"
          }
        },
        "required": [
          "client_id",
          "action_id"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
...
      "AnalyzeGameSummary": {
        "properties": {
          "blind": {
            "items": {
              "type": "string"
            },
            "title": "Blind",
            "type": "array"
          },
          "bury": {
            "items": {
              "type": "string"
            },
            "title": "Bury",
            "type": "array"
          },
          "defenderPoints": {
            "title": "Defenderpoints",
            "type": "integer"
          },
          "hands": {
            "additionalProperties": {
              "items": {
                "type": "string"
              },
              "type": "array"
            },
            "title": "Hands",
            "type": "object"
          },
          "partner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Partner"
          },
          "picker": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Picker"
          },
          "pickerPoints": {
            "title": "Pickerpoints",
            "type": "integer"
          },
          "scores": {
            "items": {
              "type": "integer"
            },
            "title": "Scores",
            "type": "array"
          }
        },
        "required": [
          "hands",
          "blind",
          "bury",
          "pickerPoints",
          "defenderPoints",
          "scores"
        ],
        "title": "AnalyzeGameSummary",
        "type": "object"
      },
...
      "AnalyzeSimulateRequest": {
        "additionalProperties": false,
        "properties": {
          "deterministic": {
            "default": false,
            "title": "Deterministic",
            "type": "boolean"
          },
          "maxSteps": {
            "default": 200,
            "title": "Maxsteps",
            "type": "integer"
          },
          "partnerMode": {
            "default": 1,
            "title": "Partnermode",
            "type": "integer"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "title": "AnalyzeSimulateRequest",
        "type": "object"
      },
      "AnalyzeSimulateResponse": {
        "properties": {
          "actionLookup": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Actionlookup",
            "type": "object"
          },
          "final": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final"
          },
          "meta": {
            "additionalProperties": true,
            "title": "Meta",
            "type": "object"
          },
          "players": {
            "items": {
              "type": "string"
            },
            "title": "Players",
            "type": "array"
          },
          "summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AnalyzeGameSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "trace": {
            "items": {
              "$ref": "#/components/schemas/AnalyzeActionDetail"
            },
            "title": "Trace",
            "type": "array"
          }
        },
        "required": [
          "meta",
          "actionLookup",
          "players",
          "trace"
        ],
        "title": "AnalyzeSimulateResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OkResponse": {
        "properties": {
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok"
        ],
        "title": "OkResponse",
        "type": "object"
      },
      "PlayerPublic": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "player_id": {
            "title": "Player Id",
            "type": "string"
          }
        },
        "required": [
          "player_id",
          "name"
        ],
        "title": "PlayerPublic",
        "type": "object"
      },
      "RedealRequest": {
        "properties": {
          "client_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Id"
          }
        },
        "title": "RedealRequest",
        "type": "object"
      },
      "RulesInput": {
        "additionalProperties": false,
        "description": "Full rules specification; unknown keys rejected.",
        "properties": {
          "doubleOnTheBump": {
            "default": true,
            "title": "Doubleonthebump",
            "type": "boolean"
          },
          "partnerMode": {
            "default": 1,
            "enum": [
              0,
              1
            ],
            "title": "Partnermode",
            "type": "integer"
          }
        },
        "title": "RulesInput",
        "type": "object"
      },
...
      "StartGameRequest": {
        "properties": {
          "client_id": {
            "title": "Client Id",
            "type": "string"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          }
        },
        "required": [
          "client_id"
        ],
        "title": "StartGameRequest",
        "type": "object"
      },
      "TablePublic": {
        "properties": {
          "fillWithAI": {
            "title": "Fillwithai",
            "type": "boolean"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "initialNames": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Initialnames",
            "type": "object"
          },
          "initialSeatOrder": {
            "items": {
              "type": "string"
            },
            "title": "Initialseatorder",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "resultsHistory": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Resultshistory",
            "type": "array"
          },
          "rules": {
            "$ref": "#/components/schemas/RulesInput"
          },
          "runningBySeat": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Runningbyseat",
            "type": "object"
          },
          "seatIsAI": {
            "additionalProperties": {
              "type": "boolean"
            },
            "title": "Seatisai",
            "type": "object"
          },
          "seatOccupants": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Seatoccupants",
            "type": "object"
          },
          "seats": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Seats",
            "type": "object"
          },
          "status": {
            "enum": [
              "open",
              "playing",
              "finished"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "rules",
          "fillWithAI",
          "seats",
          "runningBySeat",
          "seatOccupants",
          "seatIsAI",
          "host",
          "resultsHistory",
          "initialSeatOrder",
          "initialNames"
        ],
        "title": "TablePublic",
        "type": "object"
      },
      "UpdatePlayerRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "UpdatePlayerRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Sheepshead Realtime API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/analyze/simulate": {
      "post": {
        "description": "Simulate a full Sheepshead game and return detailed analysis trace.",
        "operationId": "analyze_simulate_api_analyze_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeSimulateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeSimulateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Simulate"
      }
    },
    "/api/players/{player_id}": {
      "get": {
        "operationId": "get_player_api_players__player_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Player"
      },
      "patch": {
        "operationId": "update_player_api_players__player_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "player_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Player Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlayerRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlayerPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Player"
      }
    },
...
    "/api/tables/{table_id}/action": {
      "post": {
        "operationId": "post_action_api_tables__table_id__action_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Action"
      }
    },
...
    "/api/tables/{table_id}/fill_ai": {
      "post": {
        "operationId": "fill_ai_api_tables__table_id__fill_ai_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RedealRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fill Ai"
      }
    },
...
    "/api/tables/{table_id}/redeal": {
      "post": {
        "operationId": "redeal_api_tables__table_id__redeal_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/RedealRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Redeal"
      }
    },
...
    "/api/tables/{table_id}/start": {
      "post": {
        "operationId": "start_game_api_tables__table_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TablePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Game"
      }
    },
...
  }
}